chore: moved wallet-modal
