+++ apps/website/components/organisms/wallet-modal/state.ts
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { useMutation } from 'react-query';
 import { useAccount, useSignMessage } from 'wagmi';
 
-import { useLogin } from '../../../../providers/auth/hooks';
-import { gqlAnonMethods } from '../../../../services/api';
+import { useLogin } from '../../../providers/auth/hooks';
+import { gqlAnonMethods } from '../../../services/api';
 
 export type Step =
   | 'GET_ACCOUNT'
